+++ chrome/background.js
@@ -3,25 +3,61 @@
 */
 const debug = false;
 const defaultFormat = 'markdown';
-var options; // Initialized by startProcessing in popup.js
+const extensionName = 'Copy Page Link';
 
-
-/* -------------------------------------------------------- */
-/*   Functions for extracting and processing link info      */
-/* -------------------------------------------------------- */
-
-function copyToClipboard (str) {
-  let listener = function (event) {
-    event.clipboardData.setData('text/plain', str);
-    event.preventDefault();
-    if (debug) console.log(str);
-  };
-  document.addEventListener('copy', listener);
-  document.execCommand('copy', false, null);
-  document.removeEventListener('copy', listener);
+// If lastError is undefined, return true. Otherwise, log the error
+// message to the console and return false.
+function notLastError () {
+  if (!chrome.runtime.lastError) { return true; }
+  else {
+    console.log(chrome.runtime.lastError.message);
+    return false;
+  }
 }
 
-function getFormattedLink (data) {
+(function initExtension() {
+  chrome.storage.sync.get(function (options) {
+    if (notLastError()) {
+      setTooltip(options);
+    }
+  });
+})();
+
+/* -------------------------------------------------------- */
+
+function setTooltip (options) {
+  let format = getCapitalizedFormat(options);
+  chrome.browserAction.setTitle({ title: `${extensionName}: ${format}` });
+}
+
+function getCapitalizedFormat (options) {
+  switch (options.format) {
+    case 'markdown':
+      return 'Markdown';
+    case 'html':
+      return 'HTML';
+    case 'latex':
+      return 'LaTeX';
+    case 'xml':
+      return 'XML';
+    default:
+      return 'Markdown';
+  }
+}
+
+/* -------------------------------------------------------- */
+
+//  getFormattedLink: The main function for extracting and processing
+//  the data used for creating the formatted link markup.
+
+function getFormattedLink (data, options) {
+  // Truncate selection string if length exceeds maximum
+  const maxLength = 120;
+  if (data.selection && data.selection.length > maxLength) {
+    data.selection = data.selection.substring(0, maxLength) + '…';
+  }
+
+  // Construct and return the link markup based on format option
   let name = data.selection ? data.selection : data.title;
   let format = options.format || defaultFormat;
 
@@ -45,40 +81,67 @@ function getFormattedLink (data) {
   }
 }
 
+/* ---------------------------------------------------------------- */
+
+//  processLinkData: Called by the content script or copyPageLink directly,
+//  depending on protocol of page link. First gets the extension options and
+//  calls copyToClipboard. If successful, calls the notifySuccess function.
+
 function processLinkData (data) {
-  // Truncate selection string if length exceeds maximum
-  const maxLength = 120;
-  if (data.selection && data.selection.length > maxLength) {
-    data.selection = data.selection.substring(0, maxLength) + '…';
+
+  function copyToClipboard (options) {
+    let str = getFormattedLink(data, options);
+    let listener = function (event) {
+      event.clipboardData.setData('text/plain', str);
+      event.preventDefault();
+    };
+    document.addEventListener('copy', listener);
+    document.execCommand('copy', false, null);
+    document.removeEventListener('copy', listener);
   }
-  copyToClipboard(getFormattedLink(data));
+
+  function notifySuccess (options) {
+    setTooltip(options);
+    let format = getCapitalizedFormat(options);
+    let message = `${format} formatted link copied to clipboard.`;
+
+    chrome.notifications.create({
+      "type": "basic",
+      "iconUrl": chrome.extension.getURL("copy-to-clipboard.png"),
+      "title": "Copy Page Link",
+      "message": message
+    });
+  }
+
+  // Get the options data saved in browser.storage
+  chrome.storage.sync.get(function (options) {
+    if (notLastError()) {
+      copyToClipboard(options);
+      if (notLastError()) {
+        notifySuccess(options);
+      }
+    }
+  });
 }
 
 /* ---------------------------------------------------------------- */
 
-// Because we've declared a popup for the extension, we need an entry
-// point function we can call from the popup script that replicates what
-// the browserAction.onClicked event handler would have done. That entry
-// point is the processActiveTab function.
+//  copyPageLink: The handler for the browserAction.onClicked event and thus
+//  the main entry point to the extension.
 
-const queryInfo = {currentWindow: true, active: true};
+function copyPageLink (tab) {
+  // Security policy only allows us to inject the content script that
+  // accesses title and selection for pages loaded with http or https.
+  function checkUrlProtocol (tab) {
+    return (tab.url.indexOf('http:') === 0 || tab.url.indexOf('https:') === 0);
+  }
 
-// Security policy only allows us to inject the content script that
-// accesses title and selection for pages loaded with http or https.
-
-function checkUrlProtocol (tab) {
-  return (tab.url.indexOf('http:') === 0 || tab.url.indexOf('https:') === 0);
-}
-
-function processActiveTab () {
-  chrome.tabs.query(queryInfo, function (tabs) {
-    if (checkUrlProtocol(tabs[0])) {
-      chrome.tabs.executeScript(null, { file: 'content.js' });
-    }
-    else {
-      processLinkData({ href: tabs[0].url, title: '', selection: '' });
-    }
-  });
+  if (checkUrlProtocol(tab)) {
+    chrome.tabs.executeScript(null, { file: 'content.js' });
+  }
+  else {
+    processLinkData({ href: tab.url, title: '', selection: '' });
+  }
 }
 
 /* ---------------------------------------------------------------- */
@@ -90,3 +153,7 @@ chrome.runtime.onMessage.addListener(
     processLinkData(request);
   }
 );
+
+// Listen for toolbar button activation
+
+chrome.browserAction.onClicked.addListener(copyPageLink);