+++ firefox/options.js
@@ -26,6 +26,13 @@ function setMessage (info) {
 
 browser.runtime.getPlatformInfo().then(setMessage, onError);
 
+function setTooltip (options) {
+  function callBackgroundPageFn (page) {
+    page.setTooltip(options);
+  }
+    browser.runtime.getBackgroundPage().then(callBackgroundPageFn, onError);
+}
+
 /* -------------------------------------------------------- */
 /*   Functions for saving and restoring user options        */
 /* -------------------------------------------------------- */
@@ -72,6 +79,7 @@ function saveOptions(e) {
       href: document.getElementById('href').value,
       name: document.getElementById('name').value
     };
+    setTooltip(options);
 
     browser.storage.sync.set(options).then(notifyUser, onError);
   }
@@ -89,11 +97,12 @@ function restoreOptions() {
 
     document.getElementById('msec').value = options.msec || defaultTimeout;
 
-    document.getElementById('link').value = options.link || 'link';
+    document.getElementById('link').value = options.link || 'site';
     document.getElementById('href').value = options.href || 'href';
     document.getElementById('name').value = options.name || 'name';
 
     if (debug) console.log(options);
+    setTooltip(options);
   }
 
   browser.storage.sync.get().then(setPreferences, onError);